ADD 'include/ComparePaths.awk' for path comparison Signed-off-by: Zdenek Styblik <stybla@turnovfree.net>
This script is going to be utilized when removing package from repository,
because there is no easy way to tell where is package located inside repository
and package name and its version is not enough to identify package.
So, package name and path in repository will be used instead, which should be
enough to identify package we want to remove.